Revert "Boilerplate for Lightning"
This reverts commit 818436a9
