--- application/comit_node/src/misc.rs
@@ -1,15 +1,0 @@
-use std::time::Duration;
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct Seconds(pub u64);
-
-impl From<Duration> for Seconds {
-    fn from(duration: Duration) -> Self {
-        Seconds(duration.as_secs())
-    }
-}
-
-impl From<Seconds> for Duration {
-    fn from(seconds: Seconds) -> Duration {
-        Duration::from_secs(seconds.0)
-    }
-}
