+++ application/comit_node/src/swap_protocols/ledger/ethereum.rs
@@ -1,4 +1,4 @@
-use ethereum_support::{Address, Transaction, H256};
+use ethereum_support::{Address, EtherQuantity, Transaction, H256};
 use secp256k1_support::PublicKey;
 use swap_protocols::ledger::Ledger;
 
@@ -6,6 +6,7 @@ use swap_protocols::ledger::Ledger;
 pub struct Ethereum {}
 
 impl Ledger for Ethereum {
+    type Quantity = EtherQuantity;
     type TxId = H256;
     type Pubkey = PublicKey;
     type Address = Address;
