+++ application/comit_node/src/swap_protocols/ledger/mod.rs
@@ -3,9 +3,8 @@ use std::fmt::Debug;
 
 mod bitcoin;
 mod ethereum;
-mod lightning;
 
-pub use self::{bitcoin::Bitcoin, ethereum::Ethereum, lightning::Lightning};
+pub use self::{bitcoin::Bitcoin, ethereum::Ethereum};
 use bam_api::header::{FromBamHeader, ToBamHeader};
 use http_api::ledger::{FromHttpLedger, ToHttpLedger};
 use std::hash::Hash;
@@ -18,11 +17,14 @@ pub trait Ledger:
     + 'static
     + Default
     + PartialEq
+    + Eq
+    + Hash
     + FromHttpLedger
     + ToHttpLedger
     + FromBamHeader
     + ToBamHeader
 {
+    type Quantity: Debug + Copy + DeserializeOwned + Serialize + Send + Sync + 'static;
     type TxId: Debug + Clone + DeserializeOwned + Serialize + Send + Sync + PartialEq + 'static;
     type Pubkey: Clone + Debug + Send + Sync + 'static;
     type Address: Debug + Clone + DeserializeOwned + Serialize + Send + Sync + 'static;
@@ -37,7 +39,14 @@ pub trait Ledger:
         + From<Self::Address>
         + Serialize
         + DeserializeOwned;
-    type Transaction: Debug + Clone + Send + Sync + PartialEq + 'static;
+    type Transaction: Debug
+        + Clone
+        + DeserializeOwned
+        + Serialize
+        + Send
+        + Sync
+        + PartialEq
+        + 'static;
 
     fn address_for_identity(&self, identity: Self::Identity) -> Self::Address;
 }
