+++ application/comit_node/src/swap_protocols/metadata_store.rs
@@ -10,7 +10,6 @@ pub enum RoleKind {
 pub enum LedgerKind {
     Bitcoin,
     Ethereum,
-    Lightning,
 }
 
 #[derive(Clone, Copy, Debug, Display)]
