--- application/comit_node/src/swap_protocols/rfc003/lightning/mod.rs
@@ -1,9 +1,0 @@
-use misc::Seconds;
-use secp256k1_support::PublicKey;
-use swap_protocols::{ledger::Lightning, rfc003::Ledger};
-
-impl Ledger for Lightning {
-    type LockDuration = Seconds;
-    type HtlcLocation = ();
-    type HtlcIdentity = PublicKey;
-}
