+++ application/comit_node/src/swap_protocols/rfc003/mod.rs
@@ -8,7 +8,6 @@ pub mod bob;
 pub mod ethereum;
 pub mod events;
 pub mod find_htlc_location;
-pub mod lightning;
 pub mod roles;
 pub mod state_machine;
 pub mod state_store;
